Resolvido problema na requisição, adição do script sql.

+++ src/main/java/com/example/petlezaBack/form/FormRepository.java
@@ -2,6 +2,6 @@ package com.example.petlezaBack.form;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FormRepository extends JpaRepository<Form, Long> {
+public interface FormRepository extends JpaRepository<Form, Integer> {
 
 }
